HSS connection cache now uses global process registry
Connection cache now uses global process registry instead of proplists
passed around as gen server state. Creating connection is a bottleneck
as of now and a single process creating them is untenable. This change
helps us replace a single process with a pool of processes in future.

TODO:
If there are a pool of processes creating connections, handle potential
race conditions when registering processes.