HSS Redesign
Every HSS script run request creates a Task object. Each Task object
is handled by a task process. Task process in turn creates a machine
process for every machine that is part of the target. Machine process
creates the ssh connection to the target machine and establishes the
channel and creates a channel handler process. Channel handler process
write the script output to the disk.

TODO:

* State management and propogation
* Ability to subscribe to ssh channel events (Make file writer a
  subscriber)
* Separate file writer and channel process
* Make writing to disk optional